django/requirements.txt
django
daphne
celery
SQLAlchemy
djangorestframework
aiohttp
drf-spectacular